requirements.txt
Pillow~=9.0.1
tqdm~=4.62.3
reportlab~=4.2.0